DV-211 refactor: 음성 로딩단계 수정
